Refactor: change album artwork name to MediaArtwork
